lab6_with_pg/users_service/requirements.txt
bcrypt==4.1.2
psycopg2-binary==2.9.9
fastapi==0.103.2
pymongo
requests
pyjwt
redis
